Provide a workaround for runtime operations without ARM requests
